implemented save and loadgame method

+++ src/main/java/nl/hu/cisq1/lingo/trainer/data/SpringGameRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SpringGameRepository extends JpaRepository<Game, Long> {
+public interface SpringGameRepository extends JpaRepository<Game, Integer> {
 }
